# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Exchange Rate Management System API
  version: 0.1.0
  description: >
    Exchange Rate Management System REST API. Provides exchange rate data,
    historical tracking, and AI-powered trend insights.
servers:
  - url: /api/v1
paths:
  /status:
    get:
      operationId: getStatus
      summary: Report service and database health
      tags: [Status]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceStatus'
        '503':
          description: Service is unhealthy (e.g. database unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceStatus'
  /exchange/refresh:
    post:
      operationId: refreshExchangeRates
      summary: Manually trigger an on-demand exchange rate collection run
      description: >
        Fetches the latest exchange rates from the external rate provider and upserts them,
        independent of the daily schedule. Does not affect per-currency usage counters.
      tags: [Exchange Rate Collection]
      responses:
        '200':
          description: Collection run completed; rates were fetched and upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResult'
        '502':
          description: The external rate provider was unreachable or returned an error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /exchange:
    get:
      operationId: getExchangeRate
      summary: Get the spread-adjusted exchange rate between two currencies
      tags: [Exchange Rate Lookup]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
          description: Source currency code
        - name: to
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
          description: Target currency code
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Rate date to use; omit for the most recent date both currencies share
      responses:
        '200':
          description: Spread-adjusted exchange rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateResponse'
        '400':
          description: >
            Invalid input — missing/malformed currency code, identical from/to currencies, or an
            unknown currency code
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No stored rate data for the requested (or most recent common) date
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /exchange/trend:
    get:
      operationId: getExchangeRateTrend
      summary: Get the historical spread-adjusted exchange rate series for a currency pair
      tags: [Exchange Rate Analytics]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
          description: Source currency code
        - name: to
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
          description: Target currency code
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Defaults to 29 days before today when omitted
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Defaults to today when omitted
      responses:
        '200':
          description: >
            Chronologically ordered trend points (empty array if no dates in range have data
            for both currencies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateTrendResponse'
        '400':
          description: >
            Unknown currency code, or startDate after endDate
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /exchange/usage:
    get:
      operationId: getUsageAnalytics
      summary: Get per-currency usage analytics, optionally ranked and/or recency-filtered
      tags: [Exchange Rate Usage Analytics]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Return at most this many currencies, ranked by queryCount descending
        - name: recentDays
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: >
            Only include currencies last queried within this many days of today; currencies
            never queried are excluded when this parameter is supplied
      responses:
        '200':
          description: Per-currency usage statistics (empty array if none match)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageAnalyticsResponse'
        '400':
          description: Non-positive or malformed limit/recentDays
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  schemas:
    ServiceStatus:
      type: object
      required: [status, databaseConnected, timestamp]
      properties:
        status:
          type: string
          enum: [UP, DOWN]
        databaseConnected:
          type: boolean
        timestamp:
          type: string
          format: date-time
    RefreshResult:
      type: object
      required: [currenciesCollected, rateDate]
      properties:
        currenciesCollected:
          type: integer
          format: int32
          description: Number of currencies successfully upserted in this run
        rateDate:
          type: string
          format: date
          description: The rate date reported by the provider and persisted in this run
    ProblemDetail:
      type: object
      description: RFC 9457 problem details
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    ExchangeRateResponse:
      type: object
      required: [fromCurrency, toCurrency, rate, rateDate, fromCurrencyUsageCount, toCurrencyUsageCount]
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        rate:
          type: string
          format: decimal
          description: Spread-adjusted rate, exact decimal (transported as string to avoid float precision loss)
        rateDate:
          type: string
          format: date
        fromCurrencyUsageCount:
          type: integer
          format: int64
        toCurrencyUsageCount:
          type: integer
          format: int64
    ExchangeRateTrendResponse:
      type: object
      required: [fromCurrency, toCurrency, points]
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/RateTrendPoint'
    RateTrendPoint:
      type: object
      required: [rateDate, rate]
      properties:
        rateDate:
          type: string
          format: date
        rate:
          type: string
          format: decimal
          description: Spread-adjusted rate, exact decimal (transported as string)
    UsageAnalyticsResponse:
      type: object
      required: [currencies]
      properties:
        currencies:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyUsageEntry'
    CurrencyUsageEntry:
      type: object
      required: [currencyCode, queryCount, lastQueriedAt]
      properties:
        currencyCode:
          type: string
        queryCount:
          type: integer
          format: int64
        lastQueriedAt:
          type: string
          format: date-time
          nullable: true
          description: Null if this currency has never been queried